Create own view models for each display type
