+++ Leibit.Client.WPF/Windows/Display/Views/PassengerInformation.xaml
@@ -2,8 +2,8 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:controls="clr-namespace:Leibit.Controls;assembly=Leibit.Controls.WPF">
-    
-    <Border BorderBrush="#373d6d" BorderThickness="10" Width="400" HorizontalAlignment="Left">
+
+    <Border DataContext="{Binding SelectedDisplayType.ViewModel}" BorderBrush="#373d6d" BorderThickness="10" Width="400" HorizontalAlignment="Left">
         <Grid>
             <Grid.RowDefinitions>
                 <RowDefinition Height="Auto"/>
