+++ Leibit.Client.WPF/Windows/Display/Views/PlatformDisplaySmall.xaml
@@ -2,8 +2,8 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:controls="clr-namespace:Leibit.Controls;assembly=Leibit.Controls.WPF">
-    
-    <Border BorderBrush="Black" Background="#1071df" BorderThickness="20" Width="443">
+
+    <Border DataContext="{Binding SelectedDisplayType.ViewModel}" BorderBrush="Black" Background="#1071df" BorderThickness="20" Width="443">
         <Border.Resources>
             <Style TargetType="TextBlock">
                 <Setter Property="Foreground" Value="White"/>
